Switch to gittuf GitHub app (#78)
This commit removes the gittuf specific workflows to record RSL entries
and merge attestations in favor of the app. The verification workflow is
updated to run on merges to the main branch.

NOTE: There may still be a race condition, however, between the app
recording an RSL entry for the merge and the action trying to verify it.

Signed-off-by: Aditya Sirish <aditya@saky.in>